Loading from CSV
Using D3, I made a VERY simple function to load the data, convert the needed values, and log it to the console. Although not yet very functional, we can at least get the data.

Currently facing the problem of how to access specific data from the CSV (a certain line/column). This may require a different approach than currently used.