Display data as table
We got it working to display all data (all 31k lines) in a table on the main HTML page. A button has been added to prevent the need to download the data each time the page is loaded (for the people that will work on the design of this page).

Co-Authored-By: Para-Boo <80783230+Para-Boo@users.noreply.github.com>