Can now read user input into data.
Now reads user input into data. Due to the security issues by the browser, data has to be handled in a different way prior to before when we were still hardcoding which file data should be read from. As such the current two visualisation methods currently do not function properly, but this should not be to hard to fix: not to mention we had to make some changes to how they handle data anyway.